doc/_extensions: Include doxygen's doxmlparser Python module
This module is not - hopefully - yet published on PyPi, so include its
code directly - a poor form of vendoring, if you will. Once the module
is published, this "vendoring" can go away.

Upstream for it is at
https://github.com/doxygen/doxygen/tree/master/addon/doxmlparser

There's also an open issue on doxygen repo about publishing this module:
https://github.com/doxygen/doxygen/issues/8848

It will be used on next patch to parse doxygen XML files.

Signed-off-by: Ederson de Souza <ederson.desouza@intel.com>