cmake: Pass path list in a Windows friendly way
When invoking cmake/llext-edk.cmake, CMakeLists.txt was sending the list
of includes as a colon separated list. On Windows, this list was
actually a single item, so it wasn't being parsed properly.

As colon is an expected character on Windows paths (think of C:\), this
patch changes the separator for '|', which hopefully won't be a problem
on non-Windows environments (being the pipe symbol, hopefully no-one will
use files with '|' on their names - the same could be an issue with
':', in fact).

Fixes #73069

Signed-off-by: Ederson de Souza <ederson.desouza@intel.com>