dts/arm/microchip: Add flash controller to flash0 on MEC17
The fact that flash0 node does not belong to a flash controller can
cause macro DT_MTD_FROM_FIXED_PARTITION, which gets a flash controller
for a given parition, to fail by returning the SoC node, which is not a
flash controller, as users of DT_MTD_FROM_FIXED_PARTITION would expect.

Fix this by adding a dummy flash controller which is disabled, so it
won't be used any further.

Signed-off-by: Ederson de Souza <ederson.desouza@intel.com>